[ODH-2329]: refactoring
