+++ src/test/it/resources/basic_analytics/eos_stb_reports/flat/ethernet_stats_reports/expected_result.txt
