+++ src/test/it/resources/basic_analytics/eos_stb_reports/flat/net_configuration_reports/expected_result.txt
