+++ src/test/it/resources/basic_analytics/eos_stb_reports/flat/wifi_stats_reports/configuration.yml
@@ -6,6 +6,6 @@ spark:
 #  consoleWriter: true
 
 analytics:
-  componentName: st_wifi_report
+  componentName: stb_wifi_report
   watermark: '1 minutes'
   window: '10 minutes'
