+++ src/test/it/resources/basic_analytics/eos_stb_reports/flat/wifi_stats_reports/input/wifi_reports.txt
