--- src/test/it/resources/basic_analytics/stb_general/wifi_stats_reports/expected_result.txt
@@ -1,3 +1,0 @@
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "st_wifi_report.wireless.average_downstream_kbps.rxKbps.avg", "value": 33.33}
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "st_wifi_report.wireless.average_upstream_kbps.txKbps.avg", "value": 33.33}
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "st_wifi_report.wifi_network.viewer_id.distinct_count", "value": 2}
